Update macro_unsupervised.py
(1) I have updated the cost calculation to be distance/dissimilarity. The previous code was remnant from an earlier experiment. I have also removed a redundant function in the same file, which had lesser functionality, i.e., no multipliers. 

(2) Regarding the specific dissimilarity function, during my experiments while implementing the code, I had used both distances and noticed the results to be similar. If I remember correctly, the magnitude for Rogers-Tanimoto used to be slightly higher than Jaccard-Needham distances. To let you use the function of interest, I have added the distance_type as a parameter with the default as 'Rogers-Tanimoto’, and another option as ‘Jaccard-Needham’. Feel free to use the one that seems the most apt for your dataset.